+++ tests/test_inform.py
@@ -750,3 +750,16 @@ def test_stripy():
             ack: invoked on: <date>
             {expected}
         ''').strip().format(expected=expected)
+
+
+if __name__ == '__main__':
+    # As a debugging aid allow the tests to be run on their own, outside pytest.
+    # This makes it easier to see and interpret and textual output.
+
+    defined = dict(globals())
+    for k, v in defined.items():
+        if callable(v) and k.startswith('test_'):
+            print()
+            print('Calling:', k)
+            print((len(k)+9)*'=')
+            v()